requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
gspread==6.0.2
google-auth==2.23.4
gspread-formatting==1.1.2
python-dotenv==1.0.0
playwright==1.44.0
fake-useragent==1.5.1
